Change Python SDK to use instances instead of class variables. (#8)
# Will use default config client-config.json
    client_one = Commander()

    # Will use specified client config.
    client_two = Commander(config_file_location=client-config-2.json)

Added module to mock server interactions. This should be by the
unit tests.

Added a mocking module to allow to build response message from the
secret management service. This can be used for unit tests, even in
other projects that use this SDK.

The two instances will not interfere with each other.