Add a unit test for default config handling. (#10)
Added a test to check what happens if the config is completely missing
and there is not config variables. In the past a empty JSON file would
be created and when loaded would cause an exception since a missing
value was needed. Now if missing a config file, and there are no
dictionary values, an exception is throw and a blank config file is not
created.

The second part tests a simple instantiation of Commander with no
config passed in. The default behavior will to load the config
with FileKeyValueStorage, using the default file name.

Co-authored-by: John Walstra <jwalstra@keepersecurity.com>