Fix storage problems and added more testing
Originally overriding the client key and server we only used if
no configuration object was passed in. Changed so they can be
overwritten if part of the Commander args.

If the config file is missing, create a empty JSON file instead of
a "touched" file. This will cause any place the loads the JSON
file not to break, since an empty file is not JSON.

Changed InMemoryKeyValueStorage read_storage to return a
dictionary with keys like FileKeyValueStorage does. Before it returned
nothing.

Broke up and created more test file.