Fix problem with assuming appKey exists in config
If not bound, appKey is removed from the config. During
prepare_context there was attempt to convert a None to bytes,
which threw an exception.

If app_key is None, don't try to decode it. We only use it in
unit test for knowning the secret the response should be
encrypted with.