Fix verify_ssl_certs flag, duplicate code, allow ENV to set config file.
Python bool will not convert a string to a bool value, like it does with
integers and other value. Use distutils.util.strtobool to convery the
string to a 1 or 0, then use bool() to convert to True/False.

Fix problem with server being used for default value of client_key, if
not set.

The existing_secret_key_hash check was a duplicate code block. This
affected the "else" block below. Removed the top duplicate block
of code.

Allow the default FileKeyValueStorage location to be set via an
environment value. This allows the config to before stored
in a persistent volume in a docker container. ie:

    ---
    version: '3.7'

    services:
      app:
        build: .
        environment:
          - KEEPER_SECRET_KEY=XXXXXXX
          - KEEPER_CONFIG_FILE=/config/client.json
        volumes:
          - keeper_config:/config
    volumes:
      keeper_config:
        driver: local