Move where tests directory and handle exec windows related errors.
The tests directory was inside of the module directory, which was
causing the tests to be included in the module. Moved one
directory up, which changed the tests a bit.

Some test needed to initialize the profile first. Added to test.

Tested under windows. If using a powershell script, the cmd needs
to execute 'powershell script.ps1' else an error is thrown. Adjusted
error handling to detect this and give an explanation of the error.