Keeper cli (#26)
* Add command line tool for the secret manager

When the module is installed, it will create a script called 'ksm'.
That is the CLI script.

Added the notation method to the SDK to allow simple notation to
get a value. Also add some error handling to the save method.

* Swithed to JSONPath instead of the JQ bindings.

JSONPath is pure Python while the JQ modules were binding that required
jq to be installed. This was a problem under Windows.

* Remove config file if there is an error on first connection.

If there was an error, and a config file was create, remove it. If it's
not removed, the next time it cli is run it will pick it up and complain
about not finding the client key.

* Move where tests directory and handle exec windows related errors.

The tests directory was inside of the module directory, which was
causing the tests to be included in the module. Moved one
directory up, which changed the tests a bit.

Some test needed to initialize the profile first. Added to test.

Tested under windows. If using a powershell script, the cmd needs
to execute 'powershell script.ps1' else an error is thrown. Adjusted
error handling to detect this and give an explanation of the error.

Co-authored-by: John Walstra <jwalstra@keepersecurity.com>