Removed default value for profile
Remove the default value for the --profile-name parameter.

The command default for the profile name defaulted to DEFAULT. Even
if you selected an active profile, it would always use the DEFAULT
profile.

Also renamed the method get_default_profile_name to get_active_profile_name
to make more sense. It's returning the active profile, not the DEFAULT
profile.