Getting ready for release
Moved the version from the __init__.py to the setup.py. There was problem
in both the CLI and SDK where the __init__.py was attempting to load
modules that have yet to be installed. It's a cart before the horse
problem.

Added the company header to files in CLI and SDK.

Tested installing from Pypi. In a nutshell (replace testpypi with pypi for real
server) ...

    $ python setup.py build
    $ python setup.py sdist
    $ twine upload -r testpypi -u "__token__" -p "pypi-TOKEN" --verbose dist/*

    $ pip install --extra-index-url https://test.pypi.org/simple/ keeper_sm_cli

https://test.pypi.org/project/keeper-sm-cli/

Heh ...

    $ cat << EOF > build_me.sh
    #!/bin/sh
    python setup.py build
    python setup.py sdist
    twine upload -r testpypi -u "__token__" -p "${PYPI_TOKEN}" --verbose dist/*
    EOF
    $ chmod u+x build_me.sh
    $ export PYPI_TOKEN=keeper://PYPI_RECORD_UID/custom_field/token
    $ ksm exec -- ./build_me.sh