Change the name of the config sections. (#30)
The section DEFAULT is a special INI section which is the base
for other sections. So the parser will add the values from DEFAULT
and override them with the values from the section.

The problem is when you are reading keys from sections that do not
have anything to do with SDK connection. You get all the keys.

Moved our default from 'DEFAULT' to '_default'. And changed the section
name 'common' to '_config' since it holds the config for the CLI.

Co-authored-by: John Walstra <jwalstra@keepersecurity.com>