More fixes and update from QA
KSM-153

* Reformat looks of tables
* Hide fields/custom fields that do not have any values
* Get references to other records
* Add a global exception handler to CLI. Don't show stack trace.
* Fix JSONPath problem
* Improve error message handling when saving.

The two references are addressRef and cardRef. AddressRef works, however
cardRef returns an empty folder for the response.

The JSONPath problem was the module. jsonpath-ng didn't like to do
queries where there was a conditional predicate. It would complain
about the ? in the syntax. Switched to version of the library where
this works, which is jsonpath-rw-ext.