Replace placeholder notation inline for exec (#53)
The reason for this is if a command like this ...

    export MYSECRET=keeper://XXXX/field/login
    ksm exec -- ./runme ${MYSECRET}

is run. Linux will replace the env var before we bet to it. This will
cause ./runme to get keeper://XXXX/field/login in the ARGV instead of a
secret.

The inline will parse the cmd items and replace any keeper:// notation
with values.