Fix problems with downloading value vi notation
Needed to move the output initalization to the actual output since
we don't know what kind of value we have until we want to output.

Might by str or bytes, so we can't open a file until we know that.

Added unit test for testing and fix the mock data to handle str and
byte content.