Add key rotation and fix client version (#63)
* Add key rotation and fix client version

Added a handler for http error/exceptions. It is for both
the get_secrets() and save() method. The exception handler will
format errors messages and throw the correct exception. It will
also handle the key rotation if it gets an 'key' error.

If a key error happens, the public key in the dictionary with
the server provided key_id will be used. The config is updated
with this key_id. The request that caused the excpetion will be
retried.

* Remove default test of version

It's too system specific.

* Fix typos