Add change event. Allow integration subscribe ot topic.
Created a way for integrations to subsribe to on_change topic from the SDK. For example
when key rotation occurs the intergration might not use a storage class from the SDK,
and it will need to update it's own config. When the key change request is sent from the
server, and SDK changes the public key, the event is pushed to the on_change topic. If an
integration is subscribed to the topic it's topic handler. There the integration can handle
what it needs to do.

For the CLI, the config is updated. For ansible, a notice is pushed to the log since the
config might be set from the Ansible vault which we can't modify.