Hard code the client version, for now. Add more public keys. (#74)
Hard code the client version to 16.0.1 so we can safely version the
python module until we get out of QA. Add the ECC prublic keys, 7+.
THe default is public key is is 7. Also added checks to make
sure the public key id is set in the config and it points at a
valid key.