Cli menu changes (#75)
* Add color to menus and tables.

* Windows can't use fnctl.

Needed a better way of finding the width of the windows console. Turned out
it was easier than the Linux version. Oddly using get_terminal_size on Linix
gave me a width of 80 no matter how wide I made the terminal.

* More robust version command

The problem before was if getting keeper-secrets-manager-core failed,
it won't attempt to get the CLI. They would both be Unknown.

* Handle exception when getting terminal size when not in a terminal