Move KSM logging from root loggger to a 'ksm' logger.
Did this so we can disable, change log level, etc to our logger without affecting
the root logger.

If no log_level is specified, the ksm logger will be disabled. If a log_level
is set, the ksm logger will be initialized.

    2021-08-12 15:46:01,598 | ksm | INFO | Secret key found in code
    2021-08-12 15:46:01,601 | ksm | DEBUG | Already bound
    2021-08-12 15:46:01,601 | ksm | DEBUG | DEBUG ERROR

Our logger will display 'ksm' instead of 'root'