Fix matching on mix-case types
Searching on the "type" is now case insensitive. The value
was being lowercased, however the key in the JSON wasn't. So

    pincode != pinCode

We lowercase the match value since the Web UI uses mixed case for
the label. Login != login. This caused problem for the mixcase
type text.

Custom label is still use case sensitive label search, but if the label
is missing it will switch to case insensitive type matching.