Field inflate and notation changes (#128)
* Change how fields are accessed and inflate references

Refactor code so standard and custom fields are handled the
same. Old method still exists, however created some new ones.

Prior to this, only custom fields were inflated, and only one level.
Changed the code so it can inflate values in the standard and
custom fields. It also can recurse. For example ..

Record (custom CardRef) -> Payment Card Record -> Address Record

If a custom field has a cardRef, it will pull it in, and the
paymentCard has a reference to address, and that is also pulled in.

Values are all put in one object instead of nesting them so notation
will still work.

Notation will always inflate. The CLI can turn inflation on or off.

* Add back inflate code