Update the kv parse for update, allow raw JSON to set
The prior problem was on how to split text into key/value pairs since
a key or label could contain a = character. Can't just split on the
equal character.

Since we are doing an update, we know the labels/types of the fields. So
all text should start with a known field/custom field, a =, and then a
value.  Changed to match the text to a field label/type and the equal character.
Anything after that is a value.

Now the user doesn't have to escape = chracters, and also escape the escape character.

Next added some CLI options to pass in JSON. Using the options will flag to take the value,
decode it, and then store the array or dictionary as the value.

Updated unit tests.

Bumped version.