Read version from text file
Check if a version text file exists. If it does, then read the cli
and sdk versions from that file.

This is due to the binaries not being able to detect their version. The
binary build process will create a version text file and include
it in the binary distribution.