Revert debug.
The problem was with the versions.txt files. In Linux and MacOS it was an ASCII
file. In Windows is was a utf8NoBOM encoded file. When reads as ASCII the
modules name didn't match the keys in the dictionary. Change the build process
to write an ASCII file for Windows and it works.