Version file (#202)
The binaries could not detect their version. Normally this comes from the site-packages. Instead use the GitHub action to build a versions text file containing the CLI and SDK versions. 

In the CLI code, if it detects versions.txt exists, use the versions from this file. 

The versions.txt is an ASCII encoding. This is not a problem with Linux and Mac, however the Windows will attempt to write a UTF8NoBOM encoded file unless the encoding is set to ASCII.