Added ability to block all apps
Some Windows apps will block, some will not. For example, Visual Code
is launched with code.cmd, which launches code.exe. Since code.cmd
ends quickly, we don't get a chance to edit. So we have blocking
that checks to see if the process name is actually running.

MacOS UI apps have a different kind of blocking. The application
needs to end.

Also fixed the problem with the record type json not being included
in the sdist.