Add record framework (#225)
* Add CLI change in that use the helper module

* Added ability to block all apps

Some Windows apps will block, some will not. For example, Visual Code
is launched with code.cmd, which launches code.exe. Since code.cmd
ends quickly, we don't get a chance to edit. So we have blocking
that checks to see if the process name is actually running.

MacOS UI apps have a different kind of blocking. The application
needs to end.

Also fixed the problem with the record type json not being included
in the sdist.

* Add github actions and handle parsing better

* Add test for config command

Add test for config and fixed a few problem. The color --enable/disabled did
not work.

* Allow user to recheck if record data file

If the editor they are using is nonblocking, the CLI
will complain about template markers being found before the user
saves.

This allows the user to finished editing and then typing 'r' to
recheck the file.

* Handle output and exceptions better

* Add missing module to requirements.txt and setup.py

* Tests are failing in GitHub action. Print the command results.

* Fixed unit test and error message for config commands

The unit test was failing because no keeper.ini file existed. Use the Mock
config and the Export module too make a fake keeper.ini.

Also fixed some of the error when a command fails.