Ignore false positive on notation detection (#245)
If a value of an environment variable starts with "keeper" it assumes it
is notation. If it turned out not to be notation, the code would
throw an error. The prevented any env var to have a value starting
with "keeper.""

If the SDK throws an error about "Keeper url missing information..." the CLI
now will just skip it instead of throwing an error. If it gets a different
exception message, a KsmCliException is thrown instead of the ValueError so
the CLI handles it properly (no stacktrace by default).

The notation detecting could be approved in the SDK, however this will work
for now.

Also added more unit tests.