Issue fixes (#254)
* Issue #246 keeper.ini created when using KSM_CONFIG

Abstract the keeper.ini into a Config class. If an instance of the
class was populated using env vars, the instance cannot save the
keeper.ini (does nothing).

* Fix issue #253 Error: Cannot find a record for UID all_lower_case_letters. Cannot download filename

Fix the problem of the parameter being fulling lowercased when the value, with a =, is part of
the param. Split off the value, lowercase the param, then add back the non-lowercased value.

Throw an exception when getting a secred returns no secret.

Add the file uid to the table view and allow a file to be download using the file uid instead
of only the file name. This will allow person to download duplicately named files.

* Bump version

And add check for profile name existing if making it the active profile.