Add root level method to create record from Field object list
Only from field args and from file existed, add a method to allow an
list of Field objects to be passed in to build a record. This is on the
root level, not version level.

Added a method to get a list of record types.

Fixed some of the record type schemas.