Add keeper_create and keeper_info plugins.
The `keeper_create` plugin allows the creation of a vault record from
Ansible. The result is the record uid. This plugin wil allow custom
record type if they are defined in the vars.

The `keeper_info` plugin gives information about the KSM plugins. This
is handy for debugging or getting information about the installed plugins.
For example, the ability to see what record types (including custom types)
are available.

Also shorten'd the param name for initializing a token. And removed the
hostname param since the token has that.

And figured out how to get more stdout/stderr with the test framework.