Add ability to pass in password complexity for record.
This is record wide, not per field.

This is based on the record format for the complexity. It does
include an optional param `filter_characters` which will allow
a selection of characters to be replaced in the password. For
example, '%' is an special character in PostgreSQL. Generating
a password containing it would cause a failure. The filter will
prevent a password containing '%'.