Ansible record create (#264)
* Add keeper_create and keeper_info plugins.

The `keeper_create` plugin allows the creation of a vault record from
Ansible. The result is the record uid. This plugin wil allow custom
record type if they are defined in the vars.

The `keeper_info` plugin gives information about the KSM plugins. This
is handy for debugging or getting information about the installed plugins.
For example, the ability to see what record types (including custom types)
are available.

Also shorten'd the param name for initializing a token. And removed the
hostname param since the token has that.

And figured out how to get more stdout/stderr with the test framework.

* Add keeper_password module and password complexity to keeper_create

The `keeper_password` module will return a randomly generated password. It
will allow the user to set the complexity, including excluding specific
characters.

Add this same functionality to the `keeper_create` password when autogenerating
a password.

Fixed the `keeper_redact` callback module. Replacing the regexp with just checking if
string contained secret using 'in' worked. Complex password contained a lot of characters
regex didn't like, and re.escape didn't help.

* Make config task_var test better

Due to a refactor with autogenerated configs, the test wasn't testing enough. Code worked,
test was subpar.