Fix Permission denied error when exporting profile
https://github.com/Keeper-Security/secrets-manager/issues/277

On Windows, the temp file was being opened twice which caused a
`Permission denied` error. The solution was to create a temp file,
close it, then pass the file name of the closed file. To make sure
it was removed, a `finally` statement is used to remove the temp
file, if it exists.