Fix Permission denied error when exporting profile (#279)
* Fix Permission denied error when exporting profile

https://github.com/Keeper-Security/secrets-manager/issues/277

On Windows, the temp file was being opened twice which caused a
`Permission denied` error. The solution was to create a temp file,
close it, then pass the file name of the closed file. To make sure
it was removed, a `finally` statement is used to remove the temp
file, if it exists.

* In test, config file is string not bytes.

With the changes, the prior may have been storing the config
as bytes, but now it is stored as string. The decode() in the string
was breaking it.