Fix test to run on Windows (#281)
Biggest problem was sharing an open file. Windows does not like that.
Changed tests to create an empty file, that is closed, and then
use by the tests.

The other was the script used in the exec test. It was a BASH script.
Added a BAT version for Windows.

No functional code was changed, only the test files.