If using KSM as a application, default config to "HOME" dir.
First part of this is creating a flag to indicate KSM's shell was
launched from an application wrapper. Instead if creating the config
in the working directory, it will create it in the user's "HOME"
directory.

The reason is because when KSM is launched as an application, the
current working directory may not be writeable.