Make sure mock record fields are in correct order.
Before there was a dictionary in order to prevent duplicate
fields from being added, however that really really is never going
to happen and it would be the programmers/test writer fault.

Due to a dictionary being turned into a list, there is a chance the
fields might not match the order of the record type.

Change mock to order the fields in the order they are added.

Add a test to make sure the order is correct.