Accept json config (#300)
* Accept JSON for the KSM_CONFIG environmental variable

Check to see if the Base64 begins with "{". If it does, skip
decoding the Base64 string and assume it's JSON.

The SDK accept JSON for KSM_CONFIG, the CLI didn't, Change it so
the CLI does.

This is mainly being done because to use KSM_CONFIG, with the CLI, via
a K8S secret, the JSON config would need to be Base64 encoded twice.
That would also limited where the JSON config could be used. More
functional to accept JSON in the KSM_CONFIG.

* Update the README.md changes