KSM .NET SDK and PowerShell Module: Update version to 16.6.0 and add 16.6.0 change log
The version in psd1, csproj files are upgraded to 16.6.0 across the project. This involves changing AssemblyVersion, FileVersion, PackageVersion, and ModuleVersion. The README.md is also updated to include the new changes that come with version 16.6.0, including added support for Folders, improved Passkey field type support, and support for record Transactions.