KSM-443 - Add innerFolderUid to SecretsManager.kt and RecordData.kt
Introduced an innerFolderUid to the "KeeperRecord" and "SecretsManagerResponse" classes to enable more specificity in record and folder tracking. This update was necessary for handling cases where sub-folders are in use.

Also made the 'type' property in the 'KeeperFileData' class nullable to prevent errors when a file type is not specified. This makes our service more robust when dealing with differing file types.