Update keys and data in fake_data.json
The transmission keys and data in the fake_data.json file were out of date.
This update replaces those with the current and valid ones to ensure the correct data is being used for testing and development. There was also change to the record data, this new fake data response now has up-to-date record fields