REL-4187 - Java SDK v16.6.1 - Fixed Folder Management (#506)
* KSM-443 - Add innerFolderUid to SecretsManager.kt and RecordData.kt

Introduced an innerFolderUid to the "KeeperRecord" and "SecretsManagerResponse" classes to enable more specificity in record and folder tracking. This update was necessary for handling cases where sub-folders are in use.

Also made the 'type' property in the 'KeeperFileData' class nullable to prevent errors when a file type is not specified. This makes our service more robust when dealing with differing file types.

* Update unit tests and data in fake_data.json

The transmission keys and data in the fake_data.json file were out of date.
This update replaces those with the current and valid ones to ensure the correct data is being used for testing and development. There was also change to the record data, this new fake data response now has up-to-date record fields