Azure Pipeline: Update package versions for dependencies
Bumped up the versions of dependencies in package.json and package-lock.json for "@keeper-security/secrets-manager-core" and "azure-pipelines-task-lib". The update is necessary to benefit from the fixes and features in the new versions of these packages. This includes "@keeper-security/secrets-manager-core" from "^16.3.3" to "^16.6.1", "azure-pipelines-task-lib" from "^4.0.0-preview" to "^4.4.0". Also updated "mockery" in package-lock.json from "1.7.0" to "2.1.0".