KSM-458 - Python SDK - remove core dep on helper (#519)
* KSM-440 Ansible Cache Records

Created a new Ansible action to retrieve records from the Vault and
create a encrypted serialized cache. This cache can then be set on
the attributes of the existing action. If set, the existing actions
will use the records in the cache. However, if the action uses
notation, the cache will not work. The KSM SDK notation method
needs to be changed to allow a list of Record to be passed in. Right
now it always get record from the Keeper Vault.

The new action `keeper_cache_records` takes a list of UID and/or titles.

```yaml
  tasks:
    - name: Generate a Keeper Record Cache secret
      keeper_password:
        length: 64
      register: keeper_record_cache_secret
      no_log: True

    - name: Store the Keeper Record Cache secret into variables.
      set_fact:
        keeper_record_cache_secret: "{{ keeper_record_cache_secret.password }}"
      no_log: True

    - name: Cache records. Will use keeper_record_cache_secret from above.
      keeper_cache_records:
        uids:
          - RECORD UID 1
          - RECORD UID 2
        titles:
          - TITLE 1
          - TITLE 2
      register: my_records
      no_log: True

    - name: "Get Value By UID"
      keeper_get:
        cache: "{{ my_records.cache }}"
        uid: RECORD UID 1
        field: "password"
      register: "my_password_by_uid"
```

Bonus changes

* Added the ability to select records by title to the `keeper_get`, `keeper_set`, and
  `keeper_copy` actions.
* Added a non-notation selectors for complex field values. The attributes
  `array_index` and `value_key` allow selecting which object/value in an array
  and key/value pair for dictionaries.
* Updated the test framework. Changes in Ansible broke our test framework.

* KSM-458 Remove core dep on helper module

Refactor add_custom_field not to use the helper module. The method
will accept an instance of FieldType, however it is not tied to the
keeper_secrets_manager_helper.v3.field_type.FieldType module.

Allow the ability to set the type, label, and value not using the
FieldType instance. The method will still take the param `field`,
however it will also take `field_type`, `label`, and `value`. This
allows adding a custom field without have to use the helper module.

Removed references `keeper-secrets-manager-helper` from setup.py
and requirements.txt. This should break the circular reference.

Added unit tests for the method.

---------

Co-authored-by: Max Ustinov <mustinov@keepersecurity.com>