KSM CLI: Use second element in "fields" list in secret_test.py
Changed access index from 0 to 1 while fetching "fields" list in secret_test.py, under the assertion tests. This change is due to adjustments made in the application code where the first element could possibly be None, hence to prevent False negatives and maintaining the relevance of tests, this index shift was necessary.