KSM-472 Add keeper_get_record action (#568)
Added a new action to return all the fields in a record as a dictionary.
The key in the dictionary is the field label, or field type if the label is missing.
The key is normalized to remove any non-alphanumeric characters.
Non-alphanumeric characters replaced with underscores.

To use this action must be used with the 'register' attribute in order to store
the results in memory. The results can be accessed the same way other stored
structures are stored in Ansbile.