KSM-529 Handle broken records, files and records. (#620)
Prior, if a record, file or folder had bad encryption the entire
application would break.

This change will cause bad items to be skipped. Their UID will be
logged as an error along with the exception information.

The mock classes were changed to handle creating bad items. And
unit tests created to handle bad encryption on all the items.